+++ SpringBootRestApp/src/main/java/com/bugtracker/SpringBootRestApp/dao/ProjectManagerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.bugtracker.SpringBootRestApp.model.ProjectManager;
 
 public interface ProjectManagerRepository extends CrudRepository<ProjectManager, String>{
-	ProjectManager findById(int id);
+	ProjectManager findById(Long id);
 	
 	ProjectManager findByUsername(String username);
